requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-community
langchain_groq

pytesseract 
pillow

python-dotenv`
